refactor loki client mock
