+++ pkg/httpclient/http_client.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/netobserv/network-observability-console-plugin/pkg/httpclient/httpclienttest"
 	"github.com/sirupsen/logrus"
 )
 
@@ -23,7 +22,7 @@ type httpClient struct {
 
 var slog = logrus.WithField("module", "server")
 
-func NewHTTPClient(timeout time.Duration, overrideHeaders map[string][]string, skipTLS bool, useMocks bool) Caller {
+func NewHTTPClient(timeout time.Duration, overrideHeaders map[string][]string, skipTLS bool) Caller {
 	transport := &http.Transport{
 		DialContext:     (&net.Dialer{Timeout: timeout}).DialContext,
 		IdleConnTimeout: timeout,
@@ -33,11 +32,6 @@ func NewHTTPClient(timeout time.Duration, overrideHeaders map[string][]string, s
 	if skipTLS {
 		transport.TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
 		slog.Warn("skipping TLS checks. SSL certificate verification is now disabled !")
-	}
-
-	if useMocks {
-		slog.Debug("Mocking HTTP Client")
-		return new(httpclienttest.HTTPClientJSONMock)
 	}
 
 	return &httpClient{
